feat: Make the whole project
I didn't commit a single time in between. LMAO I'M DEAD.

Made a getComputerChoice function which chooses between rock, paper
and scissors.

Made a playRound function which compares the player's
choice(made with prompt) and computer's choice and returns either
a true(player won) or false(computer won) statement and prints out
correct phrase. If there is a tie, choices get reselected and
the function calls and returns itself recursively until there is a
winner.

Made a game function which loops through five rounds(a best of five)
and prints out an overall winner at the end of the loop.